requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
bcrypt==5.0.0
Django==5.0.2
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-ninja==1.5.1
django-seed==0.3.1
ecdsa==0.19.1
Faker==38.2.0
passlib==1.7.4
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.4
toposort==1.10
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
djangorestframework
django-redis
gunicorn
django-silk